+++ mtgogui/src/appdata.rs
@@ -1,4 +1,5 @@
 pub mod paths;
+pub mod update;
 
 // Directory that stores all collection data
 pub const APP_DATA_DIR: &str = "appdata";
